Add management command for rebuilding indexes
