+++ dthm4kaiako/utils/search_utils.py
@@ -1,5 +1,8 @@
 """Search utility functions."""
 
+from django.db.models import Value
+from django.contrib.postgres.search import SearchVector
+
 
 def concat_field_values(*args):
     """Return string of field values for search indexing.
@@ -16,3 +19,25 @@ def concat_field_values(*args):
             for field in instance:
                 field_names.append(str(field))
     return ' '.join(field_names)
+
+
+def get_search_index_updater(instance):
+    """Return function for updating search index of instance."""
+    components = instance.index_contents()
+    pk = instance.pk
+
+    def on_commit():
+        search_vector_list = []
+        for weight, text in components.items():
+            search_vector_list.append(
+                SearchVector(Value(text), weight=weight)
+            )
+        search_vectors = search_vector_list[0]
+        for search_vector in search_vector_list[1:]:
+            search_vectors += search_vector
+
+        instance.__class__.objects.filter(pk=pk).update(
+            search_vector=search_vectors
+        )
+        print(f'Rebuilt index for {instance}')
+    return on_commit
